+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Support for Battery BLE Service you can use to indicate the battery level (see Example1_MockedReadings).
+
+### Changed
+- characteristicSetValue and characteristicNotify of IBLELibraryWrapper now return a bool indiciating whether the set/notify was successful
+
+
 ## [1.0.0] - 2022-07-28
 ### Added
 - Wrapper for BLE library